Exit once callback URL has been hit
This is a bit neater and prevents the state from being re-used. It also
means that we don't need to terminate and get the correct exit code. The
error will always appear in the browser and not the CLI.

I experimented with splitting out to functions, instead of inline
functions, but it didn't seem to make it much clearer at this size.

Starting the server in a goroutine means that we can't exit on bind
errors before printing the auth URL without constructing our own
listener, which doesn't seem worth the hassle.